Use post thumbnail and display feature post preview
I changed the theme to allow adding thumbnail images to posts for displaying on
the front page. Now, for posts on the front page with thumbnail images, we use
the appropriate thumbnail image. For posts without thumbnail images, we use the
same default image as before. We might want to change the default image in the
future before launching.

I also changed the main featured post to only display a portion of the content,
not the whole post. We should only show the part of the post before the `more`
tag, or things will start looking weird.